refactor: input_page: Separate widgets into files
