+++ lib/input_page.dart
@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:font_awesome_flutter/font_awesome_flutter.dart';
+import 'reusable_card.dart';
+import 'customized_cards.dart';
 
 const bottomContainerHeight = 80.0;
 const primaryCardColor = Color(0xFF1D1E33);
@@ -80,55 +82,6 @@ class _InputPageState extends State<InputPage> {
           ],
         ),
       ),
-    );
-  }
-}
-
-class GenderCards extends StatelessWidget {
-  String gender;
-  IconData genderIcon;
-
-  GenderCards({required this.gender, required this.genderIcon});
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      mainAxisAlignment: MainAxisAlignment.center,
-      children: [
-        Icon(
-          genderIcon,
-          size: 100.0,
-        ),
-        const SizedBox(
-          height: 15.0,
-        ),
-        Text(
-          gender,
-          style: const TextStyle(
-            fontSize: 20.0,
-            color: Color(0xFF8D8E98),
-          ),
-        )
-      ],
-    );
-  }
-}
-
-class ReusableCard extends StatelessWidget {
-  Color colour;
-  Widget childCard;
-
-  ReusableCard({required this.colour, required this.childCard});
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      decoration: BoxDecoration(
-        color: colour,
-        borderRadius: BorderRadius.circular(10.0),
-      ),
-      margin: const EdgeInsets.all(15.0),
-      child: childCard,
     );
   }
 }
